run all common resource tests against Non-RDF Source
Change NonRDFSourceTest to inherit from CommonResourceTest so that all
of the common tests also run against LDP Non-RDF Source. Fix any
CommonResourceTest tests so that they do not assume the content is RDF
(text/turtle).
