+++ src/main/java/org/w3/ldp/testsuite/test/CommonResourceTest.java
@@ -21,16 +21,8 @@ import org.w3.ldp.testsuite.annotations.SpecTest.METHOD;
 import org.w3.ldp.testsuite.annotations.SpecTest.STATUS;
 import org.w3.ldp.testsuite.exception.SkipMethodNotAllowedException;
 import org.w3.ldp.testsuite.http.HttpMethod;
-import org.w3.ldp.testsuite.mapper.RdfObjectMapper;
 import org.w3.ldp.testsuite.vocab.LDP;
 
-import com.hp.hpl.jena.rdf.model.Model;
-import com.hp.hpl.jena.rdf.model.ModelFactory;
-import com.hp.hpl.jena.rdf.model.Resource;
-import com.hp.hpl.jena.rdf.model.Statement;
-import com.hp.hpl.jena.rdf.model.StmtIterator;
-import com.hp.hpl.jena.util.ResourceUtils;
-import com.hp.hpl.jena.vocabulary.DCTerms;
 import com.jayway.restassured.RestAssured;
 import com.jayway.restassured.response.Response;
 import com.jayway.restassured.specification.ResponseSpecification;
@@ -115,8 +107,6 @@ public abstract class CommonResourceTest extends LdpTest {
     public void testETagHeadersGet() {
         // GET requests
         RestAssured
-            .given()
-                .header(ACCEPT, TEXT_TURTLE)
             .expect()
                 .statusCode(isSuccessful())
                 .header(ETAG, isValidEntityTag())
@@ -135,7 +125,7 @@ public abstract class CommonResourceTest extends LdpTest {
             approval = STATUS.WG_APPROVED)
     public void testETagHeadersHead() {
         // GET requests
-        RestAssured.given().header(ACCEPT, TEXT_TURTLE)
+        RestAssured
                 .expect().statusCode(isSuccessful()).header(ETAG, isValidEntityTag())
                 .when().head(getResourceUri());
     }
@@ -153,7 +143,7 @@ public abstract class CommonResourceTest extends LdpTest {
             testMethod = METHOD.AUTOMATED,
             approval = STATUS.WG_APPROVED)
     public void testLdpLinkHeader() {
-        Response response = RestAssured.given().header(ACCEPT, TEXT_TURTLE)
+        Response response = RestAssured
                 .expect().statusCode(isSuccessful())
                 .when().get(getResourceUri());
         assertTrue(
@@ -165,56 +155,6 @@ public abstract class CommonResourceTest extends LdpTest {
         );
     }
 
-    @Test(
-            groups = {MUST},
-            description = "LDP servers MUST assign the default base-URI "
-                    + "for [RFC3987] relative-URI resolution to be the HTTP "
-                    + "Request-URI when the resource already exists, and to "
-                    + "the URI of the created resource when the request results "
-                    + "in the creation of a new resource.")
-    @SpecTest(
-            specRefUri = LdpTestSuite.SPEC_URI + "#ldpr-gen-defbaseuri",
-            testMethod = METHOD.AUTOMATED,
-            approval = STATUS.WG_APPROVED)
-    public void testRelativeUriResolutionPut() {
-        skipIfMethodNotAllowed(HttpMethod.PUT);
-
-        String resourceUri = getResourceUri();
-        Response response = RestAssured
-            .given()
-                .header(ACCEPT, TEXT_TURTLE)
-            .expect()
-                .statusCode(isSuccessful())
-                .header(ETAG, isValidEntityTag())
-            .when()
-                .get(resourceUri);
-
-        String eTag = response.getHeader(ETAG);
-        Model model = response.as(Model.class, new RdfObjectMapper(resourceUri));
- 
-        // Make sure the resource is specified using a relative URI
-        ResourceUtils.renameResource(model.getResource(resourceUri), "");
-
-        // Update a property
-        updateResource(model.getResource(""));
-        
-        // Put the resource back using relative URIs.
-        Response put = RestAssured
-                .given().contentType(TEXT_TURTLE).header(IF_MATCH, eTag)
-                .body(model, new RdfObjectMapper("")) // relative URI
-                .when().put(resourceUri);
-        if (!isSuccessful().matches(put.getStatusCode())) {
-           throw new SkipException("Cannot verify relative URI resolution because the PUT request failed. Skipping test."); 
-        }
-
-        // Get the resource again to verify its content.
-        model = RestAssured.given().header(ACCEPT, TEXT_TURTLE)
-                .expect().statusCode(isSuccessful())
-                .when().get(resourceUri).as(Model.class, new RdfObjectMapper(resourceUri));
-
-        // Verify the change.
-        verifyUpdatedResource(model.getResource(resourceUri));
-    }
 
     @Test(
             enabled = false,
@@ -267,125 +207,6 @@ public abstract class CommonResourceTest extends LdpTest {
         }
 
         expectResponse.when().get(getResourceUri());
-    }
-
-    @Test(
-            groups = {MUST},
-            description = "If a HTTP PUT is accepted on an existing resource, "
-                    + "LDP servers MUST replace the entire persistent state of "
-                    + "the identified resource with the entity representation "
-                    + "in the body of the request.")
-    @SpecTest(
-            specRefUri = LdpTestSuite.SPEC_URI + "#ldpr-put-replaceall",
-            testMethod = METHOD.AUTOMATED,
-            approval = STATUS.WG_APPROVED)
-    public void testPutReplacesResource() {
-        skipIfMethodNotAllowed(HttpMethod.PUT);
-
-        if (restrictionsOnContent()) {
-            throw new SkipException("Skipping test because there are restrictions on PUT content for this resource");
-        }
-
-        // TODO: Is there a better way to test this requirement?
-        String resourceUri = getResourceUri();
-        Response response = RestAssured
-                .given()
-                        .header(ACCEPT, TEXT_TURTLE)
-                .expect()
-                        .statusCode(isSuccessful())
-                        .header(ETAG, isValidEntityTag())
-                .when()
-                        .get(resourceUri);
-
-        String eTag = response.getHeader(ETAG);
-        Model originalModel = response.as(Model.class, new RdfObjectMapper(resourceUri));
-
-        // Replace the resource with something different.
-        Model differentContent = ModelFactory.createDefaultModel();
-        final String UPDATED_TITLE = "This resources content has been replaced";
-        differentContent.add(differentContent.getResource(resourceUri),
-                DCTerms.title, UPDATED_TITLE);
-        RestAssured
-                .given().contentType(TEXT_TURTLE).header(IF_MATCH, eTag)
-                .body(differentContent, new RdfObjectMapper(resourceUri)) // relative URI
-                .expect().statusCode(isSuccessful())
-                .when().put(resourceUri);
-
-        // Get the resource again to see what's there.
-        response = RestAssured
-            .given()
-                .header(ACCEPT, TEXT_TURTLE)
-            .expect()
-                .statusCode(isSuccessful())
-                .header(ETAG, isValidEntityTag())
-            .when()
-                .get(resourceUri);
-        eTag = response.getHeader(ETAG);
-        Model updatedModel = response.as(Model.class, new RdfObjectMapper(resourceUri));
-
-        // Validate the updated resource content. The LDP server is allowed to add in
-        // some triples (for instance, dcterms:lastModified), so we can't just compare
-        // that it's exactly what we posted. Let's make sure not all of the triples
-        // from the original resource are there since we've completely replaced it,
-        // however. Also check that the title is as expected.
-        Resource updatedResource = updatedModel.getResource(resourceUri);
-        assertTrue(updatedResource.hasProperty(DCTerms.title, UPDATED_TITLE), "Expected updated resource to have title: " + UPDATED_TITLE);
-        boolean hasDifferentProperties = false;
-        Resource originalResource = originalModel.getResource(resourceUri);
-        StmtIterator iter = originalResource.listProperties();
-        while (iter.hasNext()) {
-            Statement s = iter.next();
-            if (!updatedResource.hasProperty(s.getPredicate())) {
-                hasDifferentProperties = true;
-            }
-        }
-        assertTrue(hasDifferentProperties, "The updated resource has the same properties as the original. Was it really replaced?");
-
-        // Replace the resource with its original content to clean up.
-        RestAssured
-                .given().contentType(TEXT_TURTLE).header(IF_MATCH, eTag)
-                .body(originalModel, new RdfObjectMapper(resourceUri)) // relative URI
-                .expect().statusCode(isSuccessful())
-                .when().put(resourceUri);
-    }
-
-    @Test(
-            groups = {SHOULD},
-            description = "LDP servers SHOULD allow clients to update resources "
-                    + "without requiring detailed knowledge of server-specific "
-                    + "constraints. This is a consequence of the requirement to "
-                    + "enable simple creation and modification of LDPRs.")
-    @SpecTest(
-            specRefUri = LdpTestSuite.SPEC_URI + "#ldpr-put-simpleupdate",
-            testMethod = METHOD.AUTOMATED,
-            approval = STATUS.WG_APPROVED)
-    public void testAllowUpdateResources() {
-        skipIfMethodNotAllowed(HttpMethod.PUT);
-
-        String resourceUri = getResourceUri();
-        Response response = RestAssured.given().header(ACCEPT, TEXT_TURTLE)
-                .expect().statusCode(isSuccessful()).header(ETAG, isValidEntityTag())
-                .when().get(resourceUri);
-
-        String eTag = response.getHeader(ETAG);
-        Model model = response.as(Model.class, new RdfObjectMapper(resourceUri));
-
-        // Update a property
-        updateResource(model.getResource(resourceUri));
-
-        RestAssured
-                .given().contentType(TEXT_TURTLE).header(IF_MATCH, eTag)
-                .body(model, new RdfObjectMapper(resourceUri))
-                .expect().statusCode(isSuccessful())
-                .when().put(resourceUri);
-
-        // Get the resource again to verify its content.
-        model = RestAssured.given().header(ACCEPT, TEXT_TURTLE)
-                .expect().statusCode(isSuccessful())
-                .when().get(resourceUri).as(Model.class, new RdfObjectMapper(resourceUri));
-
-        // Verify the change.
-        verifyUpdatedResource(model.getResource(resourceUri));
     }
 
     @Test(
@@ -464,22 +285,20 @@ public abstract class CommonResourceTest extends LdpTest {
         skipIfMethodNotAllowed(HttpMethod.PUT);
 
         String resourceUri = getResourceUri();
-        Model model = RestAssured
-            .given()
-                .header(ACCEPT, TEXT_TURTLE)
+        Response response = RestAssured
             .expect()
                 .statusCode(isSuccessful())
                 .header(ETAG, isValidEntityTag())
             .when()
-                .get(resourceUri).as(Model.class, new RdfObjectMapper(resourceUri));
+                .get(resourceUri);
 
         RestAssured
-                .given()
-                .contentType(TEXT_TURTLE)
-                .body(model, new RdfObjectMapper(resourceUri))
-                .expect()
+            .given()
+                .contentType(response.getContentType())
+                .body(response.asByteArray())
+            .expect()
                 .statusCode(not(isSuccessful()))
-                .when()
+            .when()
                 .put(resourceUri);
     }
 
@@ -500,23 +319,22 @@ public abstract class CommonResourceTest extends LdpTest {
         skipIfMethodNotAllowed(HttpMethod.PUT);
 
         String resourceUri = getResourceUri();
-        Model model = RestAssured
-                .given()
-                .header(ACCEPT, TEXT_TURTLE)
+        Response response = RestAssured
                 .expect()
-                .statusCode(isSuccessful()).header(ETAG, isValidEntityTag())
+                    .statusCode(isSuccessful()).header(ETAG, isValidEntityTag())
                 .when()
-                .get(resourceUri).as(Model.class, new RdfObjectMapper(resourceUri));
-        
+                    .get(resourceUri);
+        String contentType = response.getContentType();
+ 
         RestAssured
                 .given()
-                .contentType(TEXT_TURTLE)
-                .header(IF_MATCH, "These aren't the ETags you're looking for.")
-                .body(model, new RdfObjectMapper(resourceUri))
+                    .contentType(contentType)
+                    .header(IF_MATCH, "These aren't the ETags you're looking for.")
+                    .body(response.asByteArray())
                 .expect()
-                .statusCode(HttpStatus.SC_PRECONDITION_FAILED)
+                    .statusCode(HttpStatus.SC_PRECONDITION_FAILED)
                 .when()
-                .put(resourceUri);
+                    .put(resourceUri);
     }
 
     @Test(
@@ -537,24 +355,20 @@ public abstract class CommonResourceTest extends LdpTest {
 
         String resourceUri = getResourceUri();
         Response getResponse = RestAssured
-                .given()
-                .header(ACCEPT, TEXT_TURTLE)
                 .expect()
                 .statusCode(isSuccessful())
                 .header(ETAG, isValidEntityTag())
                 .when()
                 .get(resourceUri);
         
-        Model model = getResponse.as(Model.class, new RdfObjectMapper(resourceUri));
-        
         // Verify that we can successfully PUT the resource WITH an If-Match header.
         Response ifMatchResponse = RestAssured
                 .given()
-                .contentType(TEXT_TURTLE)
-                .header(IF_MATCH, getResponse.getHeader(ETAG))
-                .body(model, new RdfObjectMapper(resourceUri))
+                    .header(IF_MATCH, getResponse.getHeader(ETAG))
+                    .contentType(getResponse.contentType())
+                    .body(getResponse.asByteArray())
                 .when()
-                .put(resourceUri);
+                    .put(resourceUri);
         if (!isSuccessful().matches(ifMatchResponse.getStatusCode())) {
             throw new SkipException("Skipping test because PUT request failed with valid If-Match header.");
         }
@@ -564,10 +378,10 @@ public abstract class CommonResourceTest extends LdpTest {
         // code.
         Response noIfMatchResponse = RestAssured
                 .given()
-                .contentType(TEXT_TURTLE)
-                .body(model, new RdfObjectMapper(resourceUri))
+                    .contentType(getResponse.contentType())
+                    .body(getResponse.asByteArray())
                 .when()
-                .put(resourceUri);
+                    .put(resourceUri);
         if (isSuccessful().matches(noIfMatchResponse.getStatusCode())) {
             // It worked. This server doesn't require If-Match, which is only a
             // SHOULD requirement (see testPutRequiresIfMatch). Skip the test.
@@ -596,21 +410,20 @@ public abstract class CommonResourceTest extends LdpTest {
         skipIfMethodNotAllowed(HttpMethod.PUT);
 
         String resourceUri = getResourceUri();
-        Model model = RestAssured
-                .given()
-                .header(ACCEPT, TEXT_TURTLE)
-                .expect()
+        Response response = RestAssured
+            .expect()
                 .statusCode(isSuccessful()).header(ETAG, isValidEntityTag())
-                .when()
-                .get(resourceUri).as(Model.class, new RdfObjectMapper(resourceUri));
+            .when()
+                .get(resourceUri);
 
         RestAssured
-                .given().contentType(TEXT_TURTLE)
+            .given()
+                .contentType(response.getContentType())
                 .header(ETAG, "\"This is not the ETag you're looking for\"") // bad ETag value
-                .body(model, new RdfObjectMapper(resourceUri))
-                .expect()
+                .body(response.asByteArray())
+            .expect()
                 .statusCode(HttpStatus.SC_PRECONDITION_FAILED)
-                .when()
+            .when()
                 .put(resourceUri);
     }
 
@@ -693,45 +506,5 @@ public abstract class CommonResourceTest extends LdpTest {
         if (!supports(method)) {
             throw new SkipMethodNotAllowedException(getResourceUri(), method);
         }
-    }
-
-    // Update a resource then later test if the updates were applied (i.e., on a subsequent GET).
-    // These methods could be overwritten by subclasses.
-    private final static String TITLE_FOR_UPDATE = "LDP Test Suite: This resource has been updated... " + System.currentTimeMillis();
-
-    /**
-     * Update a resource then later test if the updates were applied (i.e., on a
-     * subsequent GET). These methods could be overwritten by subclasses.
-     *
-     * @see #verifyUpdatedResource(Resource)
-     */
-    protected void updateResource(Resource r) {
-        // Set a title.
-        r.removeAll(DCTerms.title);
-        r.addProperty(DCTerms.title, TITLE_FOR_UPDATE);
-    }
-
-    /**
-     * Update a resource then later test if the updates were applied (i.e., on a
-     * subsequent GET). These methods could be overwritten by subclasses.
-     *
-     * @see #updateResource(Resource)
-     */
-    protected void verifyUpdatedResource(Resource r) {
-        // Test the resource has the title we set.
-        assertTrue(r.hasProperty(DCTerms.title, TITLE_FOR_UPDATE), "Expected resource to have title \"" + TITLE_FOR_UPDATE + "\"");
-    }
-
-    /**
-     * Are there any restrictions on resource content? This should be true for
-     * LDP containers since PUT is not allowed to modify containment triples.
-     *
-     * @return true if there are restrictions on what triples are allowed; false
-     * if the entire resource can be replaced with any triples
-     * @see #testPutReplacesResource
-     * @see CommonContainerTest#testRejectPutModifyingContainmentTriples
-     */
-    protected boolean restrictionsOnContent() {
-        return false;
     }
 }
